pyproject.toml
[project]
name = "runlet-runners-py"
version = "0.1.0"
description = ""
authors = [
    {name = "TheAppleKingy",email = "pridanov.ali.job@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "pydantic-settings (>=2.13.0,<3.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "redis (>=7.2.0,<8.0.0)",
    "pika (>=1.3.2,<2.0.0)",
    "watchfiles (>=1.1.1,<2.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "dishka (>=1.8.0,<2.0.0)",
    "docker (>=7.1.0,<8.0.0)",
    "ploomby (>=0.1.8,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
